requirements.txt
matplotlib>=3.1.0
scipy>=1.5.2
scikit-learn>=0.23.2
torch>=1.7.1
nengo>=3.1.0
tqdm>=4.60.0
torchsummary
pandas
tensorboard
cloudpickle
neuprint-python
seaborn
networkx
ray[tune]>=1.3.0
noise>=1.2.2